requirements.txt
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
bcrypt==3.2.2
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
Bottleneck==1.3.5
Brlapi==0.8.5
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloud-init==24.4.1
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
contourpy==1.0.7
cryptography==44.0.2
cssselect==1.2.0
cupshelpers==1.0
customtkinter==5.2.2
cycler==0.11.0
darkdetect==0.8.0
dbus-python==1.3.2
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.9.0
distro-info==1.7+build1
duplicity==2.1.4
et-xmlfile==1.0.1
exceptiongroup==1.2.2
fasteners==0.18
Flask==3.1.0
fonttools==4.46.0
fs==2.4.16
github3.py==4.0.1
gyp==0.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.20.4
httpx==0.28.1
idna==3.10
iniconfig==1.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.3.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
kaggle==1.6.17
kiwisolver==0.0.0
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
lxml==5.2.1
lz4==4.0.2+dfsg
Mako==1.3.2.dev0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.6.3
mdurl==0.1.2
monotonic==1.6
mpmath==0.0.0
netaddr==0.8.0
nltk==3.8.1
numexpr==2.9.0
numpy==1.26.4
oauthlib==3.2.2
odfpy==1.4.2
olefile==0.46
openai==1.66.2
openpyxl==3.1.2
packaging==24.0
pandas==2.1.4+dfsg
paramiko==2.12.0
patsy==0.5.6
pexpect==4.9.0
pillow==10.2.0
pluggy==1.4.0
psutil==5.9.8
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pycairo==1.25.1
pycparser==2.22
pycups==2.0.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-sip==12.13.0
pyrsistent==0.20.0
pyserial==3.5
pytest==7.4.4
python-apt==2.7.7+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.49+ubuntu2
python-slugify==8.0.4
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.2
regex==2022.10.31
requests==2.32.3
rich==13.7.1
scikit-learn==1.4.1.post1
SciPy==1.11.4
seaborn==0.13.2
simplejson==3.19.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.5
sympy==1.12
systemd-python==235
tables==3.9.2
text-unidecode==1.3
threadpoolctl==3.1.0
tqdm==4.67.1
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.16.0
ufw==0.36.2
unattended-upgrades==0.1
unicodedata2==15.1.0
uritemplate==4.1.1
urllib3==2.3.0
usb-creator==0.3.16
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==3.1.3
xdg==5
xkit==0.0.0
